+++ src/polyomino-defs.hpp
@@ -38,7 +38,7 @@ struct PolyominoDef {
   const char preview_tiles[4];
 
   bool collide(Board &board, s8 row, s8 column) const;
-  void render(u8 x, u8 y) const;
+  void render(u8 x, int y) const;
   void chibi_render(u8 row, u8 column) const;
   void board_render(Board &board, s8 row, s8 column, bool jiggling) const;
 };
